Add combine method to algo results
